% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_funcs.R
\name{abs_pesp_path}
\alias{abs_pesp_path}
\title{Absolute path to PESP data folder}
\usage{
abs_pesp_path(fp_rel = NULL)
}
\arguments{
\item{fp_rel}{Optional relative path to append within the base PESP directory}
}
\value{
A character string giving the absolute path
}
\description{
Constructs an absolute path to the shared PESP data directory under the user's home directory.
Optionally appends a relative path inside the folder.
}
