% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_funcs.R
\name{clean_unknowns}
\alias{clean_unknowns}
\title{Standardize Unknown Taxon Labels}
\usage{
clean_unknowns(df, std_sp, std_suffix)
}
\arguments{
\item{df}{Dataframe containing a \code{Taxon} column}

\item{std_sp}{Logical; if \code{TRUE}, convert all \code{spp.} variants to \code{sp.} (default: \code{TRUE})}

\item{std_suffix}{Logical; if \code{TRUE}, remove trailing descriptors (e.g., \verb{sp. A}, \code{sp. (B)}) (default: \code{TRUE})}
}
\value{
Dataframe with cleaned \code{Taxon} values and a \code{log} attribute listing changed entries
}
\description{
Cleans and standardizes taxon names by normalizing "unknown" variants,
simplifying species suffixes, and optionally converting all \code{spp.} to \code{sp.}.

Specifically:
\itemize{
\item Replaces case-insensitive variants of "unknown", "unidentified", or "undetermined" with \code{"Unknown"}
\item Converts trailing \verb{sp. X} or \verb{spp. X} to just \code{sp.} or \code{spp.}
\item Optionally converts all \code{spp.} to \code{sp.} for consistency
\item Convert \verb{cf. Unknown x} to \verb{Unknown x}
}
}
