% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_geom.R
\docType{data}
\name{delta_geom}
\alias{delta_geom}
\title{Delta subregion geometry (Mahardja)}
\format{
An object of class sf with regional boundaries
}
\source{
deltamapr::R_EDSM_Subregions_Mahardja
}
\usage{
delta_geom
}
\description{
Regional boundary designations for the Sacramento–San Joaquin Delta.
}
\keyword{datasets}
