% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_funcs.R
\name{flag_outliers}
\alias{flag_outliers}
\title{Flag statistical outliers in a measurement column}
\usage{
flag_outliers(
  df,
  col,
  station_col = Station,
  cutoff = 3,
  add_flag = FALSE,
  show_plot = TRUE
)
}
\arguments{
\item{df}{input dataframe containing the specified measurement column and
\code{QualityCheck}, \code{Notes}, \code{Date}, \code{Station}, and \code{Taxon} columns}

\item{col}{unquoted column name of the measurement variable to evaluate
(must be one of \code{Cells_per_mL}, \code{Units_per_mL}, or
\code{Biovolume_per_mL})}

\item{cutoff}{numeric value for the absolute robust z-score threshold
used to identify outliers (default = 3)}

\item{show_plot}{logical; if \code{TRUE}, prints a scatter plot with
flagged outliers in red (default = TRUE)}
}
\value{
the input dataframe with updated \code{QualityCheck} values and
an updated \code{log} attribute containing a dataframe of flagged rows
}
\description{
Identifies outliers in log-transformed measurement data using robust z-scores
(median and MAD), flags them in the \code{QualityCheck} column, and optionally
generates a scatter plot showing flagged points. The function also appends a
dataframe of flagged rows to the \code{log} attribute.
}
