% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_funcs.R
\name{from_meta}
\alias{from_meta}
\title{Add metadata values to main dataframe by Date}
\usage{
from_meta(df, df_meta, column)
}
\arguments{
\item{df}{the main dataframe with a 'Date' column}

\item{df_meta}{A metadata dataframe with 'Starting Date' and optionally 'Ending Date' columns,
plus one or more columns to join}

\item{column}{A string specifying the name of the column in \code{df_meta} to join into \code{df}.}
}
\value{
A dataframe that contains the original \code{df} with additional columns from \code{df_meta} joined by matching date ranges
}
\description{
Expands date ranges from the metadata dataframe, then joins selected columns from that metadata
into the main dataframe based on matching dates.
}
