% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyto_typos.R
\docType{data}
\name{phyto_typos}
\alias{phyto_typos}
\title{Phytoplankton taxon name typo lookup table}
\format{
A dataframe with columns:
\describe{
\item{Taxon}{The misspelled or non-standard taxon name as it appears in raw data}
\item{TaxonCorrected}{The corrected taxon name to substitute}
}
}
\usage{
phyto_typos
}
\description{
A reference table of known misspellings and non-standard taxon name variants
used to correct raw data prior to synonym resolution. Created by DWR staff.
Used internally by \code{correct_taxon_typos()}.
}
\keyword{datasets}
