% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_funcs.R
\name{read_phyto_taxa}
\alias{read_phyto_taxa}
\title{Read-in phyto taxonomy key list}
\usage{
read_phyto_taxa()
}
\value{
A dataframe of filtered metadata
}
\description{
Read in phytoplankton taxonomy key list (based on AlgaeBase; Algal Groups from Tiffany Brown at DWR).
}
