% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_funcs.R
\name{remove_nodata}
\alias{remove_nodata}
\title{Remove rows with no density data}
\usage{
remove_nodata(
  df,
  density_cols = c("Cells_per_mL", "Units_per_mL", "Biovolume_per_mL")
)
}
\arguments{
\item{df}{A dataframe containing density measurement columns}

\item{density_cols}{A character vector of column names to check for zero or NA values.
Defaults to c('Cells_per_mL', 'Units_per_mL', 'Biovolume_per_mL')}
}
\value{
A dataframe with zeros replaced by NA and rows removed where all specified
columns are NA.
}
\description{
Replaces zeros with NA in the specified density columns and
removes rows where all of those columns are NA.
}
