% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa_funcs.R
\name{calc_data_bsa}
\alias{calc_data_bsa}
\title{calc reporting units}
\usage{
calc_data_bsa(
  df,
  unit_col = "Unit Abundance",
  cell_col = "Total Number of Cells",
  calc_cols = c("Units", "Cells", "Biovolume")
)
}
\description{
calc reporting units
}
