% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_funcs.R
\name{get_edi_file}
\alias{get_edi_file}
\title{Download and read specific files from an EDI data package}
\usage{
get_edi_file(pkg_id, fname)
}
\arguments{
\item{pkg_id}{The EDI package ID (e.g., "1017")}

\item{fnames}{A character vector of filenames (or fragments) to match against package entities}
}
\value{
A named list of dataframes, where each name corresponds to a matched filename
}
\description{
Downloads specified files by name from the latest revision of an EDI data package,
and reads them into a named list of dataframes.
}
