% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_funcs.R
\name{read_meta_file}
\alias{read_meta_file}
\title{Read-in metadata file for programs}
\usage{
read_meta_file(program_name)
}
\arguments{
\item{program_name}{Name of the program to filter by}
}
\value{
A dataframe of filtered metadata with no missing ending dates
}
\description{
Reads in the PESP metadata Excel file and filters it to include only rows for the specified program.
Any missing \verb{Ending Date} values are replaced with the current system date.
}
